chore: rename fixture file
